+++ bonsai-ai/tests/ws.py
@@ -1,12 +1,22 @@
 # Copyright (C) 2018 Bonsai, Inc.
 
 import json
+import logging
 import sys
 from tornado import web, websocket, ioloop
 from google.protobuf.json_format import Parse
 
 from bonsai_ai.proto.generator_simulator_api_pb2 import ServerToSimulator
 from bonsai_ai.proto.generator_simulator_api_pb2 import SimulatorToServer
+
+
+def count_me(fnc):
+    """ decorator to count function calls in a class variable
+    """
+    def increment(self, *args, **kwargs):
+        type(self)._count += 1
+        return fnc(self, *args, **kwargs)
+    return increment
 
 
 class BrainRequestHandler(web.RequestHandler):
@@ -51,6 +61,9 @@ class BonsaiWS(websocket.WebSocketHandler):
     '''
     _PROTOCOL_FILE = "proto_bin/cartpole_wire.json"
     _PREDICT = False
+    _FLAKY = False
+    _fail_point = 10
+    _fail_duration = 8
 
     _prev = ServerToSimulator.UNKNOWN
     _dispatch = {
@@ -91,6 +104,13 @@ class BonsaiWS(websocket.WebSocketHandler):
 
     _message_data = None
 
+    # count connection attemts *AND* incoming messages to
+    # control flaky-mode behavior.
+    # rejects incoming connections while
+    # _fail_point <= _count < _fail_point + _fail_duration
+
+    _count = 0
+
     def _dispatch_mtype(self, incoming):
         if self._PREDICT:
             return self._dispatch_pred.get(self._prev, {}).get(
@@ -107,6 +127,16 @@ class BonsaiWS(websocket.WebSocketHandler):
             raise Exception(
                 "Sim ID missing from SimulatorToServer message!")
 
+    @count_me
+    def get(self, *args, **kwargs):
+        if self._FLAKY and \
+           self._count > self._fail_point and \
+           self._count < self._fail_point + self._fail_duration:
+            self.set_status(503)
+            return
+
+        return super(BonsaiWS, self).get(*args, **kwargs)
+
     def open(self):
         ''' When the first client connects, read the message data from
         disk'''
@@ -116,11 +146,18 @@ class BonsaiWS(websocket.WebSocketHandler):
             j = f.read()
             self._message_data = json.loads(j)
 
+    @count_me
     def on_message(self, in_bytes):
         from_sim = SimulatorToServer()
         from_sim.ParseFromString(in_bytes)
         self._validate_message(from_sim)
         mtype = self._dispatch_mtype(from_sim.message_type)
+
+        if self._FLAKY and \
+           self._count > self._fail_point and \
+           self._count < self._fail_point + self._fail_duration:
+            self.close(code=1008, reason="Policy violation.")
+            return
 
         # dict->json->Message->binary
         msg_dict = self._message_data[mtype]
@@ -158,6 +195,23 @@ def set_predict_mode(predict):
     BonsaiWS._PREDICT = predict
 
 
+def set_flaky_mode(flaky):
+    BonsaiWS._FLAKY = flaky
+
+
+def set_fail_duration(duration):
+    duration = int(duration)
+    if duration < 0:
+        raise ValueError(
+            'Fail duration cannot be negative. '
+            'If you wish to turn off flakiness, use set_flaky_mode.')
+    BonsaiWS._fail_duration = duration
+
+
+def reset_count():
+    BonsaiWS._count = 0
+
+
 def close_bonsai_ws():
     loop = ioloop.IOLoop.current()
     loop.add_callback(loop.stop)
@@ -169,6 +223,10 @@ if __name__ == "__main__":
         open_bonsai_ws(sys.argv[1])
         if sys.argv[2] == "P":
             set_predict_mode(True)
+        elif sys.argv[2] == "F":
+            set_flaky_mode(True)
+
+        logging.getLogger('tornado.access').disabled = True
         ioloop.IOLoop.current().start()
     except KeyboardInterrupt:
         close_bonsai_ws()